Add article helper to remove logic from article view
