+++ app/helpers/application_helper.rb
@@ -1,15 +1,2 @@
 module ApplicationHelper
-
-  def featured_owner(article)
-    article.first.user.Name
-  end
-
-  def featured_title(article)
-    article.first.Title
-  end
-
-  def featured_content(article)
-    article.first.Text
-  end
-
 end
