+++ app/helpers/article_helper.rb
@@ -1,0 +1,15 @@
+module ArticleHelper
+
+  def featured_owner(article)
+    article.first.user.Name
+  end
+
+  def featured_title(article)
+    article.first.Title
+  end
+
+  def featured_content(article)
+    article.first.Text
+  end
+
+end
